fixed regresssion in aggregators
