+++ lib/aggregator/index.js
@@ -39,25 +39,25 @@ class Aggregator {
 
 	print (style) {
 		const postProcessors = [];
-		let queries = this.queries;
+
 		this.sharedFunctions.forEach(funcConf => {
 			if (postProcessing.isPostProcessor(funcConf.func)) {
 				postProcessors.push(funcConf);
 			} else {
-				queries = queries.map(kq => kq.inheritMethod(funcConf.func, funcConf.params));
+				this.queries = this.queries.map(kq => kq[funcConf.func].apply(kq, funcConf.params));
 			}
 		});
 		style = style || this._printer;
 
 		if (style === 'qs' || style === 'qo' || style === 'url') {
-			return Promise.all(queries.map(q => q.print(style)));
+			return Promise.all(this.queries.map(q => q.print(style)));
 		}
 
 		if (this._table) {
 			return Promise.resolve(printers.call(this, style))
 		}
 
-		return Promise.all(queries.map(q => q.print('json')))
+		return Promise.all(this.queries.map(q => q.print('json')))
 			.then(() => this.aggregator())
 			.then(() => {
 				postProcessors.forEach(opts => {
@@ -112,7 +112,7 @@ class Aggregator {
 		aggregator.aggregator = this.aggregator;
 		aggregator.aggregatorName = this.aggregatorName;
 		aggregator._printer = this._printer;
-		aggregator._postProcessors = this._postProcessors && this._postProcessors.slice().map(f => Object.assign({}, f));
+		aggregator.sharedFunctions = this.sharedFunctions && this.sharedFunctions.slice().map(f => Object.assign({}, f));
 		if (withData) {
 			aggregator._table = this.getTable() && this.getTable().clone();
 		}
