+++ lib/keen-query/index.js
@@ -433,13 +433,6 @@ class KeenQuery {
 		return tabulate.apply(this, [].slice.call(arguments))
 	}
 
-	inheritMethod (name, params) {
-		//TODO maybe use this for toString() at some point
-		const kq = this[name].apply(this._getInstance(), params || []);
-		return kq;
-	}
-
-
 	getTable () {
 		return this._table;
 	}
