+++ package.json
@@ -29,7 +29,7 @@
   "homepage": "https://github.com/Financial-Times/keen-query#readme",
   "devDependencies": {
     "chai": "^3.4.1",
-    "fetch-mock": "^4.3.1",
+    "fetch-mock": "^4.4.0",
     "mocha": "^2.4.2",
     "next-build-tools": "^5.16.1"
   }
